app/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

aiohttp==3.9.5
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
opencv-python==4.9.0.80
paho-mqtt==2.0.0
pymongo==4.7.2
python-dotenv==1.0.1
